chore: Remove logic from static class to make it more testable.
